* feat: implement scene loader

* add intersections with basic objects (#6)

* Feat/basic object intersections (#7)

* add intersections with basic objects

* add calculations of triangle normals by barycentric coordinates. fix box normals calculation

* feat: add cameras (#8)

* Implement materials/lights (#9)

* feat: add some code

* fix: huge bugfixes

* fix: huge bugfixes, cleanup

* add example

(cherry picked from commit 2b1683c67fd6146204a4715887719b6f858560bb)

* Update README.md

(cherry picked from commit a30477693ae124c3717bd219fcd8af29530f6bc4)

* Update README.md

+++ src/ComputerGraphics.Jpeg/JpegEncoderInternal.cs
@@ -244,7 +244,7 @@ namespace ComputerGraphics.Jpeg
             {
                 bitWriter.Write(id);
                 bitWriter.Write(0x11);
-                bitWriter.Write(id == 1 ? 0 : 1);
+                bitWriter.Write((byte) (id == 1 ? 0 : 1));
             }
 
             WriteLuminanceAndChrominance(bitWriter);
@@ -262,7 +262,7 @@ namespace ComputerGraphics.Jpeg
             {
                 bitWriter.Write(id);
 
-                bitWriter.Write(id == 1 ? 0x00 : 0x11);
+                bitWriter.Write((byte) (id == 1 ? 0x00 : 0x11));
             }
 
             bitWriter.Write(0);
